docker/cpp_sandbox/requirements.txt
Django>=3.2
djangorestframework>=3.12
psutil>=5.8
